Move new functionalities and structs into swap_bundle.rs file
