+++ tests/zsa.rs
@@ -1,31 +1,25 @@
 mod builder;
 
 use crate::builder::{verify_action_group, verify_bundle, verify_swap_bundle};
-use bridgetree::BridgeTree;
-use incrementalmerkletree::Hashable;
-use orchard::bundle::{ActionGroupAuthorized, Authorized, SwapBundle};
-use orchard::issuance::{verify_issue_bundle, IssueBundle, IssueInfo, Signed, Unauthorized};
-use orchard::keys::{IssuanceAuthorizingKey, IssuanceValidatingKey};
-use orchard::note::{AssetBase, ExtractedNoteCommitment};
 
-use orchard::tree::{MerkleHashOrchard, MerklePath};
 use orchard::{
     builder::{Builder, BundleType},
-    bundle::Authorized,
+    bundle::{ActionGroupAuthorized, Authorized},
     circuit::{ProvingKey, VerifyingKey},
-    issuance::{
-        compute_asset_desc_hash, verify_issue_bundle, AwaitingNullifier, IssueBundle, IssueInfo,
-        Signed,
-    },
-    issuance_auth::{IssueAuthKey, IssueValidatingKey, ZSASchnorr},
+    issuance::{verify_issue_bundle, IssueBundle, IssueInfo, Signed, Unauthorized},
     keys::{FullViewingKey, PreparedIncomingViewingKey, Scope, SpendAuthorizingKey, SpendingKey},
-    note::{AssetBase, ExtractedNoteCommitment, Nullifier},
+    keys::{IssuanceAuthorizingKey, IssuanceValidatingKey},
+    note::{AssetBase, ExtractedNoteCommitment},
+    note_encryption::OrchardDomain,
     orchard_flavor::OrchardZSA,
-    primitives::OrchardDomain,
+    swap_bundle::SwapBundle,
     tree::{MerkleHashOrchard, MerklePath},
     value::NoteValue,
     Address, Anchor, Bundle, Note,
 };
+
+use bridgetree::BridgeTree;
+use incrementalmerkletree::Hashable;
 use rand::rngs::OsRng;
 use shardtree::{store::memory::MemoryShardStore, ShardTree};
 use std::collections::HashSet;
